requirements.txt
absl-py==0.11.0
airsim==1.3.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.5
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gym==0.17.3
h5py==2.10.0
idna==2.10
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
numpy==1.19.3
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow @ file:///C:/tmp/tensorflow-2.4.0rc3-cp38-cp38-win_amd64.whl
tensorflow-estimator==2.4.0rc0
termcolor==1.1.0
tf-agents-nightly==0.7.0.dev20201201
tfp-nightly==0.12.0.dev20201202
tornado==4.5.3
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
